Factorize a lot of code in navigation panel strategies
