+++ Backoffice/NavigationPanel/Strategies/AdministrationPanelStrategy.php
@@ -21,25 +21,7 @@ class AdministrationPanelStrategy extends AbstractNavigationPanelStrategy
     const ROLE_ACCESS_SITE = 'ROLE_ACCESS_SITE';
     const ROLE_ACCESS_LOG = 'ROLE_ACCESS_LOG';
 
-    protected $name;
-    protected $role;
-    protected $weight;
-    protected $parent;
     protected $template;
-
-    /**
-     * @param string $name
-     * @param string $role
-     * @param int    $weight
-     * @param string $parent
-     */
-    public function __construct($name, $role, $weight = 0, $parent)
-    {
-        $this->name = $name;
-        $this->role = $role;
-        $this->weight = $weight;
-        $this->parent = $parent;
-    }
 
     /**
      * @return string
@@ -47,38 +29,6 @@ class AdministrationPanelStrategy extends AbstractNavigationPanelStrategy
     public function show()
     {
         return $this->render($this->getTemplate());
-    }
-
-    /**
-     * @return string
-     */
-    public function getName()
-    {
-        return $this->name;
-    }
-
-    /**
-     * @return string
-     */
-    public function getParent()
-    {
-        return $this->parent;
-    }
-
-    /**
-     * @return int
-     */
-    public function getWeight()
-    {
-        return $this->weight;
-    }
-
-    /**
-     * @return string
-     */
-    public function getRole()
-    {
-        return $this->role;
     }
 
     /**
